chore: move docker-binaries-extractor to the inputs library
the inputs library is in charge of getting out the data that we need for later analysis.
moving the docker binaries extractor over there and adding a temporary placeholder for the static binaries extractor.
