+++ lib/inputs/binaries/docker/node.ts
@@ -1,6 +1,6 @@
 import { valid } from "semver";
-import { Docker, DockerOptions } from "../../docker";
-import { Binary } from "../types";
+import { Binary } from "../../../analyzer/types";
+import { Docker, DockerOptions } from "../../../docker";
 
 export { extract, installedByPackageManager };
