+++ lib/inputs/binaries/static/index.ts
@@ -1,0 +1,1 @@
+// TODO
